requirements.txt
rdflib==6.2.0
SPARQLWrapper==2.0.0
requests==2.28.0
python-dotenv==1.0.0
